// rhdeservice/clients/e-portal/src/app/shared/model/request/request-by-oid.ts
import * as uuid from 'uuid';
import { APPLICATION_NAME, CLIENT, VERSION } from '../../constant/constant';
import { SearchParams } from './search-params';

export class Header {
    requestId = uuid.v4();
    requestClient = CLIENT;
    requestType: string;
    requestSource = APPLICATION_NAME;
    requestSourceService: string;
    requestVersion = VERSION;
    requestTime: Date = new Date();
    requestTimeoutInSeconds = 20;
    requestRetryCount = 0;
}
class Body {
    oid: String;
}

export class RequestByOid {
    header: Header = new Header();
    meta: Object = new Object();
    body: Body = new Body();
}